move softfloat tests to their own file
