+++ pydrofoil/test/test_supportcode.py
@@ -8,7 +8,7 @@ from pydrofoil.bitvector import Integer, SmallInteger, BigInteger, MININT, Spars
 from pydrofoil.bitvector import GenericBitVector, SmallBitVector
 from hypothesis import given, strategies, assume, example, settings
 
-from rpython.rlib.rarithmetic import r_uint, intmask, r_ulonglong
+from rpython.rlib.rarithmetic import r_uint, intmask
 from rpython.rlib.rbigint import rbigint
 
 MAXINT = sys.maxint
@@ -1028,296 +1028,6 @@ def test_pow2():
     # check that small results use small ints
     for i in range(63):
         assert supportcode.pow2_i(None, i).val == 2 ** i
-
-# softfloat
-
-class DummyMachine(object): pass
-
-
-def test_softfloat_f16add():
-    machine = DummyMachine()
-    supportcode.softfloat_f16add(machine, 0, 0b0011110000000000, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 0b0011111000000000
-
-def test_softfloat_f16div():
-    machine = DummyMachine()
-    supportcode.softfloat_f16div(machine, 0, 0b0011110000000000, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 0b0100000000000000
-
-def test_softfloat_f16eq():
-    machine = DummyMachine()
-    supportcode.softfloat_f16eq(machine, 0b0011100000000000, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 1
-
-def test_softfloat_f16le():
-    machine = DummyMachine()
-    supportcode.softfloat_f16le(machine, 0b0011100000000000, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 1
-
-def test_softfloat_f16lt():
-    machine = DummyMachine()
-    supportcode.softfloat_f16lt(machine, 0b0011100000000000, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 0
-
-def test_softfloat_f16mul():
-    machine = DummyMachine()
-    supportcode.softfloat_f16mul(machine, 0, 0b0011110000000000, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 0b0011100000000000
-
-def test_softfloat_f16muladd():
-    machine = DummyMachine()
-    supportcode.softfloat_f16muladd(machine, 0, 0b0011110000000000, 0b0011100000000000, 0b0011110000000000)
-    assert machine._reg_zfloat_result == 0b0011111000000000
-
-def test_softfloat_f16sqrt():
-    machine = DummyMachine()
-    supportcode.softfloat_f16sqrt(machine, 0, 0b0100010000000000)
-    assert machine._reg_zfloat_result == 0b0100000000000000
-
-def test_softfloat_f16sub():
-    machine = DummyMachine()
-    supportcode.softfloat_f16sub(machine, 0, 0b0011110000000000, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 0b0011100000000000
-
-def test_softfloat_f32add():
-    machine = DummyMachine()
-    supportcode.softfloat_f32add(machine, 0, 0b00111111000000000000000000000000, 0b00111111100000000000000000000000)
-    assert machine._reg_zfloat_result == 0b00111111110000000000000000000000
-
-def test_softfloat_f32div():
-    machine = DummyMachine()
-    supportcode.softfloat_f32div(machine, 0, 0b00111111100000000000000000000000, 0b01000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b00111111000000000000000000000000
-
-def test_softfloat_f32eq():
-    machine = DummyMachine()
-    supportcode.softfloat_f32eq(machine, 0b00111111000000000000000000000000, 0b00111111000000000000000000000000)
-    assert machine._reg_zfloat_result == 1
-
-def test_softfloat_f32le():
-    machine = DummyMachine()
-    supportcode.softfloat_f32le(machine, 0b00111111000000000000000000000000, 0b00111111000000000000000000000000)
-    assert machine._reg_zfloat_result == 1
-
-def test_softfloat_f32lt():
-    machine = DummyMachine()
-    supportcode.softfloat_f32lt(machine, 0b00111111000000000000000000000000, 0b00111111000000000000000000000000)
-    assert machine._reg_zfloat_result == 0
-
-def test_softfloat_f32mul():
-    machine = DummyMachine()
-    supportcode.softfloat_f32mul(machine, 0, 0b00111111100000000000000000000000, 0b01000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b01000000000000000000000000000000
-
-def test_softfloat_f32muladd():
-    machine = DummyMachine()
-    supportcode.softfloat_f32muladd(machine, 0, 0b00111111100000000000000000000000, 0b00111111000000000000000000000000, 0b00111111100000000000000000000000)
-    assert machine._reg_zfloat_result == 0b00111111110000000000000000000000
-
-def test_softfloat_f32sqrt():
-    machine = DummyMachine()
-    supportcode.softfloat_f32sqrt(machine, 0, 0b01000000100000000000000000000000)
-    assert machine._reg_zfloat_result == 0b01000000000000000000000000000000
-
-def test_softfloat_f32sub():
-    machine = DummyMachine()
-    supportcode.softfloat_f32sub(machine, 0, 0b01000000000000000000000000000000, 0b00111111100000000000000000000000)
-    assert machine._reg_zfloat_result == 0b00111111100000000000000000000000
-
-def test_softfloat_f64add():
-    machine = DummyMachine()
-    supportcode.softfloat_f64add(machine, 0, 0b0011111111110000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0011111111111000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f64div():
-    machine = DummyMachine()
-    supportcode.softfloat_f64div(machine, 0, 0b0011111111110000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0100000000000000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f64eq():
-    machine = DummyMachine()
-    supportcode.softfloat_f64eq(machine, 0b0011111111100000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 1
-
-def test_softfloat_f64le():
-    machine = DummyMachine()
-    supportcode.softfloat_f64le(machine, 0b0011111111100000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 1
-
-def test_softfloat_f64lt():
-    machine = DummyMachine()
-    supportcode.softfloat_f64lt(machine, 0b0011111111100000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0
-
-def test_softfloat_f64mul():
-    machine = DummyMachine()
-    supportcode.softfloat_f64mul(machine, 0, 0b0011111111110000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0011111111100000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f64muladd():
-    machine = DummyMachine()
-    supportcode.softfloat_f64muladd(machine, 0, 0b0011111111110000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000, 0b0011111111110000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0011111111111000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f64sqrt():
-    machine = DummyMachine()
-    supportcode.softfloat_f64sqrt(machine, 0, 0b0100000000010000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0100000000000000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f64sub():
-    machine = DummyMachine()
-    supportcode.softfloat_f64sub(machine, 0, 0b0011111111110000000000000000000000000000000000000000000000000000, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0011111111100000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f16tof32():
-    machine = DummyMachine()
-    supportcode.softfloat_f16tof32(machine, 0, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 0b00111111000000000000000000000000
-
-def test_softfloat_f32tof16():
-    machine = DummyMachine()
-    supportcode.softfloat_f32tof16(machine, 0, 0b00111111000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0011100000000000
-
-def test_softfloat_f16tof64():
-    machine = DummyMachine()
-    supportcode.softfloat_f16tof64(machine, 0, 0b0011100000000000)
-    assert machine._reg_zfloat_result == 0b0011111111100000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f64tof16():
-    machine = DummyMachine()
-    supportcode.softfloat_f64tof16(machine, 0, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0011100000000000
-
-def test_softfloat_f32tof64():
-    machine = DummyMachine()
-    supportcode.softfloat_f32tof64(machine, 0, 0b00111111000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0011111111100000000000000000000000000000000000000000000000000000
-
-def test_softfloat_f64tof32():
-    machine = DummyMachine()
-    supportcode.softfloat_f64tof32(machine, 0, 0b0011111111100000000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b00111111000000000000000000000000
-
-def test_softfloat_f16toi32():
-    machine = DummyMachine()
-    supportcode.softfloat_f16toi32(machine, 0, 0b1100010100000000)
-    assert machine._reg_zfloat_result == 0b11111111111111111111111111111011
-
-def test_softfloat_f16toi64():
-    machine = DummyMachine()
-    supportcode.softfloat_f16toi64(machine, 0, 0b1100010100000000)
-    assert machine._reg_zfloat_result == 0b1111111111111111111111111111111111111111111111111111111111111011
-
-def test_softfloat_f16toui32():
-    machine = DummyMachine()
-    supportcode.softfloat_f16toui32(machine, 0, 0b0100010100000000)
-    assert machine._reg_zfloat_result == 0b00000000000000000000000000000101
-
-def test_softfloat_f16toui64():
-    machine = DummyMachine()
-    supportcode.softfloat_f16toui64(machine, 0, 0b0100010100000000)
-    assert machine._reg_zfloat_result == 0b0000000000000000000000000000000000000000000000000000000000000101
-
-def test_softfloat_f32toi32():
-    machine = DummyMachine()
-    supportcode.softfloat_f32toi32(machine, 0, 0b11000000101000000000000000000000)
-    assert machine._reg_zfloat_result == 0b11111111111111111111111111111011
-
-def test_softfloat_f32toi64():
-    machine = DummyMachine()
-    supportcode.softfloat_f32toi64(machine, 0, 0b11000000101000000000000000000000)
-    assert machine._reg_zfloat_result == 0b1111111111111111111111111111111111111111111111111111111111111011
-
-def test_softfloat_f32toui32():
-    machine = DummyMachine()
-    supportcode.softfloat_f32toui32(machine, 0, 0b01000000101000000000000000000000)
-    assert machine._reg_zfloat_result == 0b00000000000000000000000000000101
-
-def test_softfloat_f32toui64():
-    machine = DummyMachine()
-    supportcode.softfloat_f32toui64(machine, 0, 0b01000000101000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0000000000000000000000000000000000000000000000000000000000000101
-
-def test_softfloat_f64toi32():
-    machine = DummyMachine()
-    supportcode.softfloat_f64toi32(machine, 0, r_ulonglong(0b1100000000010100000000000000000000000000000000000000000000000000))
-    assert machine._reg_zfloat_result == 0b11111111111111111111111111111011
-
-def test_softfloat_f64toi64():
-    machine = DummyMachine()
-    supportcode.softfloat_f64toi64(machine, 0, r_ulonglong(0b1100000000010100000000000000000000000000000000000000000000000000))
-    assert machine._reg_zfloat_result == 0b1111111111111111111111111111111111111111111111111111111111111011
-
-def test_softfloat_f64toui32():
-    machine = DummyMachine()
-    supportcode.softfloat_f64toui32(machine, 0, 0b0100000000010100000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b00000000000000000000000000000101
-
-def test_softfloat_f64toui64():
-    machine = DummyMachine()
-    supportcode.softfloat_f64toui64(machine, 0, 0b0100000000010100000000000000000000000000000000000000000000000000)
-    assert machine._reg_zfloat_result == 0b0000000000000000000000000000000000000000000000000000000000000101
-
-def test_softfloat_i32tof16():
-    machine = DummyMachine()
-    supportcode.softfloat_i32tof16(machine, 0, 0b11111111111111111111111111111011)
-    assert machine._reg_zfloat_result == 0b1100010100000000
-
-def test_softfloat_i32tof32():
-    machine = DummyMachine()
-    supportcode.softfloat_i32tof32(machine, 0, 0b11111111111111111111111111111011)
-    assert machine._reg_zfloat_result == 0b11000000101000000000000000000000
-
-def test_softfloat_i32tof64():
-    machine = DummyMachine()
-    supportcode.softfloat_i32tof64(machine, 0, 0b11111111111111111111111111111011)
-    assert machine._reg_zfloat_result == 0b1100000000010100000000000000000000000000000000000000000000000000
-
-def test_softfloat_i64tof16():
-    machine = DummyMachine()
-    supportcode.softfloat_i64tof16(machine, 0, r_ulonglong(0b1111111111111111111111111111111111111111111111111111111111111011))
-    assert machine._reg_zfloat_result == 0b1100010100000000
-
-def test_softfloat_i64tof32():
-    machine = DummyMachine()
-    supportcode.softfloat_i64tof32(machine, 0, r_ulonglong(0b1111111111111111111111111111111111111111111111111111111111111011))
-    assert machine._reg_zfloat_result == 0b11000000101000000000000000000000
-
-def test_softfloat_i64tof64():
-    machine = DummyMachine()
-    supportcode.softfloat_i64tof64(machine, 0, r_ulonglong(0b1111111111111111111111111111111111111111111111111111111111111011))
-    assert machine._reg_zfloat_result == 0b1100000000010100000000000000000000000000000000000000000000000000
-
-def test_softfloat_ui32tof16():
-    machine = DummyMachine()
-    supportcode.softfloat_ui32tof16(machine, 0, 0b00000000000000000000000000000101)
-    assert machine._reg_zfloat_result == 0b0100010100000000
-
-def test_softfloat_ui32tof32():
-    machine = DummyMachine()
-    supportcode.softfloat_ui32tof32(machine, 0, 0b00000000000000000000000000000101)
-    assert machine._reg_zfloat_result == 0b01000000101000000000000000000000
-
-def test_softfloat_ui32tof64():
-    machine = DummyMachine()
-    supportcode.softfloat_ui32tof64(machine, 0, 0b00000000000000000000000000000101)
-    assert machine._reg_zfloat_result == 0b0100000000010100000000000000000000000000000000000000000000000000
-
-def test_softfloat_ui64tof16():
-    machine = DummyMachine()
-    supportcode.softfloat_ui64tof16(machine, 0, 0b0000000000000000000000000000000000000000000000000000000000000101)
-    assert machine._reg_zfloat_result == 0b0100010100000000
-
-def test_softfloat_ui64tof32():
-    machine = DummyMachine()
-    supportcode.softfloat_ui64tof32(machine, 0, 0b0000000000000000000000000000000000000000000000000000000000000101)
-    assert machine._reg_zfloat_result == 0b01000000101000000000000000000000
-
-def test_softfloat_ui64tof64():
-    machine = DummyMachine()
-    supportcode.softfloat_ui64tof64(machine, 0, 0b0000000000000000000000000000000000000000000000000000000000000101)
-    assert machine._reg_zfloat_result == 0b0100000000010100000000000000000000000000000000000000000000000000
 
 
 # memory
